pyproject.toml
[tool.poetry]
name = "dl-lecture-competition-pub"
version = "0.1.0"
description = ""
authors = ["nishino-k4"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.0"
torch = "^2.3.1"
torchmetrics = "^1.4.0.post0"
hydra-core = "^1.3.2"
wandb = "^0.17.4"
termcolor = "^2.4.0"
tqdm = "^4.66.4"
einops = "^0.8.0"
scipy = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
